Move to avoid circular dependencies.
