+++ chemclipse/plugins/org.eclipse.chemclipse.support.ui/META-INF/MANIFEST.MF
@@ -36,7 +36,6 @@ Export-Package: org.eclipse.chemclipse.support.ui.activator,
  org.eclipse.chemclipse.support.ui.preferences,
  org.eclipse.chemclipse.support.ui.preferences.editors,
  org.eclipse.chemclipse.support.ui.preferences.fieldeditors,
- org.eclipse.chemclipse.support.ui.processors,
  org.eclipse.chemclipse.support.ui.provider,
  org.eclipse.chemclipse.support.ui.richtext,
  org.eclipse.chemclipse.support.ui.services,
