+++ chemclipse/plugins/org.eclipse.chemclipse.xxd.process.ui/META-INF/MANIFEST.MF
@@ -11,10 +11,14 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.chemclipse.support.ui;bundle-version="0.8.0",
  org.eclipse.chemclipse.xxd.process;bundle-version="0.8.0",
  org.eclipse.chemclipse.support;bundle-version="0.8.0",
- org.eclipse.e4.core.di
+ org.eclipse.e4.core.di,
+ org.eclipse.chemclipse.processing;bundle-version="0.9.0",
+ org.eclipse.chemclipse.logging;bundle-version="0.9.0",
+ org.eclipse.chemclipse.rcp.ui.icons;bundle-version="0.9.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-21
 Bundle-ActivationPolicy: lazy
-Export-Package: org.eclipse.chemclipse.xxd.process.ui.menu,
+Export-Package: org.eclipse.chemclipse.xxd.process,
+ org.eclipse.chemclipse.xxd.process.ui.menu,
  org.eclipse.chemclipse.xxd.process.ui.preferences
 Import-Package: org.eclipse.chemclipse.chromatogram.xxd.report.preferences,
  org.eclipse.chemclipse.converter.preferences
