+++ chemclipse/plugins/org.eclipse.chemclipse.xxd.process.ui/src/org/eclipse/chemclipse/xxd/process/l10n/messages.properties
@@ -1,0 +1,57 @@
+add=Add
+allowedRange=Allowed range (MIN = {0})(MAX = {1})
+allowedRetentionTimeRange=Allowed retention time range: {0} - {1} minutes
+clear=Clear
+clipboardTableDefaultSorting=Clipboard Table Default Sorting
+copyClipboard=Copy to Clipboard
+clipboardSettings=Clipboard Settings
+copyHeaderToClipboard=Copy Header To Clipboard
+copyHeaderToolTip=Copy the header on export.
+valueDelimiterToolTip=Select the value delimiter.
+noClipboardSettingsAvailable=The table clipboard settings are not available.
+tableColumns=Table Columns
+deselectAll=Deselect All
+enterIon=Enter an ion.
+errorMessageOddIncludingZero=Has to be an odd number including zero.
+errorMessageOdd=Has to be an odd number.
+errorMessageEven=Has to be an even number.
+errorMessageRange=Number outside of range {0} - {1}
+exportTableSelection=Export Table Selection
+fileName=File Name
+filePath=File Path
+files=Files
+inputMustBeIntegerOrIntegerRange=The input must be an integer value or an integer value range.
+inputMustBeValidRange=The input must be a valid range.
+ions=Ions
+ionValueAlreadyExists=The ion value exists already.
+labelCategory=Category
+labelDataType=Data Type
+labelDescription=Description
+labelID=ID
+labellabelNA=n.a.
+labelName=Name
+maximumAllowedUndoSteps=Maximum allowed undo steps
+processorImage=Processor Image
+remove=Remove
+removeAllFilesFromList=Remove all files from the list?
+removeTICbeforeAddingIonValues=The TIC value 0 still exists. Remove it before adding ion values.
+removeTICbeforeAddingWavelengthValues=the TIC value 0 still exists. Remove it before adding wavelength values.
+searchAvailableProcessorItems=Search the available processor items.
+selectAll=Select All
+selectEntriesinList=Please select one or more entries in the list.
+selectFile=Select File
+settings=Settings
+showLabelsInToolbar=Show Labels in Toolbar
+showText=Show Text
+standardValuesWaterNitrogenSolventTailingColumnBleed=Standard values are 18 (water), 28 (nitrogen), 84 (solvent tailing), 207 (column bleed).
+supportSettings=Support settings
+tablePopUpMenu=Table PopUp Menu
+ticMustNotBeAdded=The TIC value (0) must not be added. If the list is empty, 0 is the default.
+toolbar=Toolbar
+columns=Columns
+resetColumnOrder=Reset Column Order (Close Required)
+resetColumnWidth=Reset Column Width (Close Required)
+wavelengths=Wavelengths
+wavelengthValueAlreadyExists=The wavelength value exists already.
+standardWavelengthRanges=Approximate UV Range from 100 to 400 nm, approximate visible range from 400 to 800 nm.
+enterWavelength=Enter a Wavelength.
