+++ chemclipse/plugins/org.eclipse.chemclipse.xxd.process.ui/src/org/eclipse/chemclipse/xxd/process/l10n/messages_de.properties
@@ -1,0 +1,55 @@
+add=Hinzufügen
+allowedRange=Erlaubter Bereich (MIN = {0})(MAX = {1})
+allowedRetentionTimeRange=Erlaubter Retentionszeitbereich: {0} - {1} Minuten
+clear=Löschen
+clipboardTableDefaultSorting=Standardsortierung der Zwischenablagetabelle
+copyClipboard=In die Zwischenablage kopieren
+clipboardSettings=Einstellungen Zwischenablage
+copyHeaderToClipboard=Kopfzeilen in die Zwischenablage kopieren
+copyHeaderToolTip=Kopfzeilen beim Export schreiben
+valueDelimiterToolTip=Trennzeichen
+noClipboardSettingsAvailable=Die Einstellungen der Zwischenablage stehen nicht zur Verfügung.
+tableColumns=Tabellenspalten
+deselectAll=Alle abwählen
+enterIon=Ion eingeben.
+errorMessageOddIncludingZero=Zahl muss ungerade sein inklusive null.
+errorMessageOdd=Zahl muss ungerade sein.
+errorMessageEven=Zahl muss gerade sein.
+fileName=Dateiname
+filePath=Dateipfad
+files=Dateien
+inputMustBeIntegerOrIntegerRange=Die Eingabe muss ganzzahlig oder eine ganzzahliger Wertebereich sein.
+inputMustBeValidRange=Die Eingabe muss ein gültiger Bereich sein.
+ions=Ionen
+ionValueAlreadyExists=Der Ionenwert existiert bereits.
+labelCategory=Kategorie
+labelDataType=Datentyp
+labelDescription=Beschreibung
+labelID=ID
+labellabelNA=n.v.
+labelName=Name
+maximumAllowedUndoSteps=maximale Rückgängigkeitsschritte
+processorImage=Prozessorbild
+remove=Entferne
+removeAllFilesFromList=Entferne alle Dateien aus der Liste?
+removeTICbeforeAddingIonValues=Der TIC Wert (0) existiert weiterhin und muss vor dem Hinzufügen weiterer Ionen entfernt werden.
+removeTICbeforeAddingWavelengthValues=Der TIC Wert (0) existiert weiterhin und muss vor dem Hinzufügen weiterer Wellenlängen entfernt werden.
+searchAvailableProcessorItems=Durchsuche alle verfügbaren Prozessoren.
+selectAll=alles auswählen
+selectEntriesinList=Bitte ein oder mehrere Einträge der Liste auswählen.
+selectFile=Datei auswählen
+settings=Einstellungen
+showLabelsInToolbar=Zeige Beschriftungen in der Symbolleiste.
+showText=Zeige Text
+standardValuesWaterNitrogenSolventTailingColumnBleed=Standardwerte sind 18 (Wasser), 28 (Stickstoff), 84 (Lösungsmittelrückstände), 207 (Säulenbluten).
+supportSettings=Unterstützungs-Einstellungen
+tablePopUpMenu=Tabellen-PopUp-Menü
+ticMustNotBeAdded=Der Wert für TIC (0) darf nicht hinzugefügt werden. Eine leere Liste entspricht 0 (TIC).
+toolbar=Symbolleiste
+columns=Spalten
+resetColumnOrder=Spaltenreihenfolge zurücksetzen (Schliessen erforderlich)
+resetColumnWidth=Spaltenbreite zurücksetzen (Schliessen erforderlich)
+wavelengths=Wellenlängen
+wavelengthValueAlreadyExists=Die Wellenlänge ist bereits in der Liste vorhanden.
+standardWavelengthRanges=Ultraviolettstrahlung ungefähr von 100 bis 400 nm, sichtbares Licht ungefähr von 400 bis 800 nm.
+enterWavelength=Wellenlänge eingeben.
